refactor: LottoNumbers.java 삭제
- 불필요 도메인이라 판단
