+++ src/test/java/lotto/entity/WinningTextsTest.java
@@ -2,7 +2,6 @@ package lotto.entity;
 
 import org.junit.jupiter.api.Test;
 
-import java.util.List;
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
